Moving navbar components to dedicated folder, switching navbar links to buttons instead of <a/> tags
