+++ src/app/layout.tsx
@@ -1,9 +1,9 @@
 import type { Metadata } from "next";
 import { use } from "react";
 import { Inter } from "next/font/google";
-import Navbar from "./components/navbar";
-import "./globals.css";
+import { Navbar } from "./components/navbar/navbar";
 import { getMetadata } from "./utilities/config-provider";
+import "./globals.css";
 
 const inter = Inter({ subsets: ["latin"] });
