+++ src/app/static/config/nav.json
@@ -14,6 +14,12 @@
           "type": "link",
           "url": "https://bellinghamtenantsunion.org/",
           "target": "_blank"
+        },
+        {
+          "name": "Residential Landlord-Tenant Act",
+          "type": "link",
+          "url": "https://apps.leg.wa.gov/rcw/default.aspx?cite=59.18",
+          "target": "_blank"
         }
       ]
     },
